Sending messages which are shown is now possible.
The client handler adds the possibility to display messages.
These messages are send via DBus to this handler from the client.
Registering and Unregistering was reworked. If a client unregisters
its corresponding dbus path is deleted. Furthermore mutli client support
was added. This enable the user to add multiple clients to one client
handler.