Use different debugging format if running as systemd service
The pyNoter server executable has a new option which will cause the
logger to use a different format. The intention for this is, that
systemd already adds time and date in front of each log entry. Hence,
having it too in the format is useless.

Adapt the service file accordingly.