Rework revise method so that it fits to the behavior of the notification library
The revise method previously reported a message which has the append
flag set, but a different subject as the referenced message, as a
revising message. However, the notification library will only combine
messages which have the same subject. Hence, in this example the message
should be reported as not revising. Which is now fixed.